+++ apps/backend/src/modules/categories/category.repository.ts
@@ -1,0 +1,54 @@
+import { type Repository } from "~/libs/types/types.js";
+
+import { CategoryEntity } from "./category.entity.js";
+import { type CategoryModel } from "./category.model.js";
+
+class CategoryRepository implements Repository {
+	private categoryModel: typeof CategoryModel;
+
+	public constructor(categoryModel: typeof CategoryModel) {
+		this.categoryModel = categoryModel;
+	}
+
+	public async create(entity: CategoryEntity): Promise<CategoryEntity> {
+		const { key, name } = entity.toNewObject();
+
+		const routeCategory = await this.categoryModel
+			.query()
+			.insert({ key, name })
+			.returning("*")
+			.execute();
+
+		return CategoryEntity.initialize(routeCategory);
+	}
+
+	public async findAll(): Promise<CategoryEntity[]> {
+		const routeCategories = await this.categoryModel.query().execute();
+
+		return routeCategories.map((routeCategory) =>
+			CategoryEntity.initialize(routeCategory),
+		);
+	}
+
+	public async findByKey(key: string): Promise<CategoryEntity | null> {
+		const routeCategory = await this.categoryModel
+			.query()
+			.whereRaw("LOWER(key) LIKE ?", [`%${key.toLowerCase()}%`])
+			.first()
+			.execute();
+
+		return routeCategory ? CategoryEntity.initialize(routeCategory) : null;
+	}
+
+	public async findByName(name: string): Promise<CategoryEntity | null> {
+		const routeCategory = await this.categoryModel
+			.query()
+			.whereRaw("LOWER(name) LIKE ?", [`%${name.toLowerCase()}%`])
+			.first()
+			.execute();
+
+		return routeCategory ? CategoryEntity.initialize(routeCategory) : null;
+	}
+}
+
+export { CategoryRepository };