+++ apps/backend/src/modules/points-of-interest/points-of-interest.service.test.ts
@@ -52,12 +52,7 @@ describe("PointsOfInterestService", () => {
 			name: mockPointOfInterest.name,
 		});
 
-		assert.deepStrictEqual(result, {
-			description: mockPointOfInterest.description,
-			id: mockPointOfInterest.id,
-			location: mockPointOfInterest.location,
-			name: mockPointOfInterest.name,
-		});
+		assert.deepStrictEqual(result, pointOfInterestEntity.toObject());
 	});
 
 	it("findAll should return all points of interest", async () => {
@@ -200,5 +195,74 @@ describe("PointsOfInterestService", () => {
 		const result = await pointsOfInterestService.delete(EXISTING_ID);
 
 		assert.strictEqual(result, true);
+	});
+
+	it("findPaginated should return paginated points of interest", async () => {
+		const mockEntities = [
+			PointsOfInterestEntity.initialize({
+				createdAt: "2025-08-14T00:00:00Z",
+				description: "Point Of Interest Test Description 1",
+				id: 1,
+				location: {
+					coordinates: TEST_COORDINATES,
+					type: "Point" as const,
+				},
+				name: "Point 1",
+				updatedAt: "2025-08-14T00:00:00Z",
+			}),
+			PointsOfInterestEntity.initialize({
+				createdAt: "2025-08-15T00:00:00Z",
+				description: "Point Of Interest Test Description 2",
+				id: 2,
+				location: {
+					coordinates: TEST_COORDINATES,
+					type: "Point" as const,
+				},
+				name: "Point 2",
+				updatedAt: "2025-08-15T00:00:00Z",
+			}),
+			PointsOfInterestEntity.initialize({
+				createdAt: "2025-08-16T00:00:00Z",
+				description: "Point Of Interest Test Description 3",
+				id: 3,
+				location: {
+					coordinates: TEST_COORDINATES,
+					type: "Point" as const,
+				},
+				name: "Point 3",
+				updatedAt: "2025-08-16T00:00:00Z",
+			}),
+		];
+
+		const mockTotal = mockEntities.length;
+		const totalPages = Math.ceil(mockTotal / 10);
+
+		const pointsOfInterestRepository = {
+			findPaginated: (() =>
+				Promise.resolve({
+					items: mockEntities,
+					total: mockTotal,
+				})) as PointsOfInterestRepository["findPaginated"],
+		} as PointsOfInterestRepository;
+
+		const pointsOfInterestService = new PointsOfInterestService(
+			pointsOfInterestRepository,
+		);
+
+		const result = await pointsOfInterestService.findPaginated({
+			page: 1,
+			perPage: 10,
+			search: undefined,
+		});
+
+		const expectedData = mockEntities.map((item) => item.toSummaryObject());
+		assert.deepStrictEqual(result.data, expectedData);
+
+		assert.deepStrictEqual(result.meta, {
+			currentPage: 1,
+			itemsPerPage: 10,
+			total: mockTotal,
+			totalPages,
+		});
 	});
 });