+++ apps/backend/src/modules/routes/route.model.ts
@@ -1,11 +1,25 @@
 import { Model } from "objection";
 
 import { DatabaseTableName } from "~/libs/modules/database/database.js";
+import { type LineStringGeometry } from "~/libs/types/types.js";
 
+import { CategoryModel } from "../categories/category.model.js";
 import { PointsOfInterestModel } from "../points-of-interest/points-of-interest.model.js";
 
 class RouteModel extends Model {
 	public static readonly relationMappings = {
+		categories: {
+			join: {
+				from: "routes.id",
+				through: {
+					from: "route_categories.route_id",
+					to: "route_categories.category_id",
+				},
+				to: "categories.id",
+			},
+			modelClass: CategoryModel,
+			relation: Model.ManyToManyRelation,
+		},
 		pois: {
 			join: {
 				from: "routes.id",
@@ -25,7 +39,11 @@ class RouteModel extends Model {
 		return DatabaseTableName.ROUTES;
 	}
 
+	public createdByUserId!: number;
 	public description!: string;
+	public distance!: number;
+	public duration!: number;
+	public geometry!: LineStringGeometry;
 	public id!: number;
 	public name!: string;
 	public pois!: {