+++ apps/frontend/src/libs/components/app-layout/libs/helpers/check-path-ends-with-route.helper.ts
@@ -1,0 +1,5 @@
+const checkPathEndsWithRoute = (routes: string[], path: string): boolean => {
+	return routes.some((route) => path.endsWith(route));
+};
+
+export { checkPathEndsWithRoute };