+++ apps/frontend/src/libs/components/app-layout/libs/types/types.ts
@@ -1,0 +1,1 @@
+export { type RouteList } from "./route-list.type.js";