+++ apps/frontend/src/modules/points-of-interest/points-of-interest-api.ts
@@ -6,6 +6,8 @@ import { type APIResponse } from "~/libs/types/types.js";
 
 import { PointsOfInterestApiPath } from "./libs/enums/enums.js";
 import {
+	type PointsOfInterestPaginatedResponseDto,
+	type PointsOfInterestQueryRequest,
 	type PointsOfInterestRequestDto,
 	type PointsOfInterestResponseDto,
 } from "./libs/types/types.js";
@@ -35,6 +37,22 @@ class PointOfInterestApi extends BaseHTTPApi {
 		);
 
 		return (await response.json()) as APIResponse<PointsOfInterestResponseDto>;
+	}
+
+	public async findPaginated(
+		payload: PointsOfInterestQueryRequest,
+	): Promise<APIResponse<PointsOfInterestPaginatedResponseDto>> {
+		const response = await this.load(
+			this.getFullEndpoint(PointsOfInterestApiPath.ROOT, {}),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "GET",
+				query: payload,
+			},
+		);
+
+		return (await response.json()) as APIResponse<PointsOfInterestPaginatedResponseDto>;
 	}
 
 	public async getById(