+++ packages/shared/src/modules/points-of-interest/libs/validation-schemas/validation-schemas.ts
@@ -1,3 +1,3 @@
-export { pointOfInterestCreate } from "./point-of-interest-create.validation-schema.js";
-export { pointOfInterestUpdate } from "./point-of-interest-update.validation-schema.js";
-export { pointsOfInterestSearchQuery } from "./points-of-interest-search-query.validation-schema.js";
+export { pointsOfInterestCreate } from "./points-of-interest-create.validation-schema.js";
+export { pointsOfInterestQuery } from "./points-of-interest-query.validation-schema.js";
+export { pointOfInterestUpdate } from "./points-of-interest-update.validation-schema.js";